feat: instead of copy component IconButton One document story for both the buttons
