+++ src/components/buttons/button.stories.mdx
@@ -1,0 +1,36 @@
+import { Canvas, Meta, Story } from '@storybook/addon-docs';
+import { IconButton } from './icon-button';
+import { Button } from './button';
+import { Regular, WithIcon } from './button.stories';
+
+<Meta title="Components/Button" component={(Button, IconButton)} />
+
+# Button
+
+Those components offer multiple styled Buttons for different situations. They all take a `onClick` property to control its functionality.
+Beyond this property the single components differ a little bit in use.
+
+### Overview of Buttons
+
+---
+
+#### Regular
+
+This is the Regular Button, which should be used in all the standard use cases. Pass the `label` as `children` to the Button.
+
+<Canvas>
+  <Story story={Regular} />
+</Canvas>
+
+#### IconButton
+
+This is the Icon Button. It is always of the size `24x24`. You can choose the displayed Icon by passing an <a href="/docs/components-icon--regular">Icon</a>.
+
+When passing an Icon to the Button, you have to:
+
+- use the default size `IconSize.NORMAL`
+- choose your wanted color of the icon by wrapping it in a Container with that color as font color
+
+<Canvas>
+  <Story story={WithIcon} />
+</Canvas>
