--- src/components/buttons/copy.stories.mdx
@@ -1,15 +1,0 @@
-import { Canvas, Meta, Story } from '@storybook/addon-docs';
-import { Copy } from './copy';
-import { Regular } from './copy.stories';
-
-<Meta title="Components/Copy" component={Copy} />
-
-# Copy
-
-A Button/Copy Component can be used to copy a specific text which is passes through the `textToBeCopied` property.
-
-#### In Action
-
-<Canvas>
-  <Story story={Regular} />
-</Canvas>
