--- src/components/buttons/copy.stories.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { ComponentStory, ComponentMeta } from '@storybook/react';
-import { Copy } from './copy';
-
-const Template: ComponentStory<typeof Copy> = (args) => <Copy {...args} />;
-
-export const Regular = Template.bind({});
-Regular.args = {
-  textToBeCopied: 'Nothing at all!!!',
-};
